# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Inventory Management System API",
        "description": "API for managing devices inventory in a company",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/devices": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Отримати всі девайси",
                "description": "Отримати всі девайси з database",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Devices"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Добавити новий девайс",
                "description": "Створити новий девайс",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddDevice"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad JSON data"
                    }
                }
            }
        },
        "/devices/{id}": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Знайти девайс за ID",
                "description": "Знайти девайс за айді",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Device ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Device"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Devices"
                ],
                "summary": "Відредагувати девайс за ID",
                "description": "Update device by ID in the database",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Device ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddDevice"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Device"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "400": {
                        "description": "Bad JSON data"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Devices"
                ],
                "summary": "Видалити девайс за ID",
                "description": "Delete device by ID from the database",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Device ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/devices/{id}/assign": {
            "put": {
                "tags": [
                    "Devices"
                ],
                "summary": "Призначити девайс до User",
                "description": "Assign device to a user by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Device ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignDevice"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Device"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "400": {
                        "description": "Already assigned"
                    }
                }
            }
        },
        "/devices/{id}/unassign": {
            "put": {
                "tags": [
                    "Devices"
                ],
                "summary": "Unassign Device from User",
                "description": "Unassign device from a user by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Device ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Device"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "400": {
                        "description": "Already unassigned"
                    }
                }
            }
        },
        "/devices/{id}/image": {
            "put": {
                "tags": [
                    "Devices"
                ],
                "summary": "Upload Device Image",
                "description": "Upload device image by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Device ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "400": {
                        "description": "Bad Image"
                    }
                }
            },
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Retrieve Device Image",
                "description": "Retrieve device image by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Device ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "image/png": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Retrieve All Users",
                "description": "Retrieve all users from the database",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Users"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Add New User",
                "description": "Add a new user to the database",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad JSON data"
                    },
                    "409": {
                        "description": "User already exists"
                    }
                }
            }
        },
        "/users/{username}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Retrieve User by Username",
                "description": "Retrieve user by username from the database",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "User username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/users/{username}/devices": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Retrieve User Devices",
                "description": "Retrieve user devices by username from the database",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "User username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Devices"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Device": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "desc": {
                        "type": "string"
                    },
                    "serial_number": {
                        "type": "string"
                    },
                    "manufacturer": {
                        "type": "string"
                    },
                    "assigned_to": {
                        "type": "string"
                    },
                    "image_path": {
                        "type": "string"
                    }
                }
            },
            "AddDevice": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Enter the name of the device."
                    },
                    "desc": {
                        "type": "string",
                        "description": "Enter the description of the device."
                    },
                    "serial_number": {
                        "type": "string",
                        "description": "Enter the serial number of the device."
                    },
                    "manufacturer": {
                        "type": "string",
                        "description": "Enter the manufacturer of the device."
                    }
                }
            },
            "Devices": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/Device"
                }
            },
            "AssignDevice": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    }
                }
            },
            "Users": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/User"
                }
            }
        }
    }
}